requirements.txt
alembic==1.8.1
anyio==3.6.2
apispec==6.0.2
apispec-webframeworks==0.5.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
camelcase==0.2
certifi==2022.9.24
click==8.1.3
colorama==0.4.5
distlib==0.3.6
Django==4.1.2
djangorestframework==3.14.0
fastapi==0.87.0
FastAPI-SQLAlchemy==0.2.1
filelock==3.8.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pipenv==2022.11.11
platformdirs==2.5.4
psycopg2==2.9.3
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
sqlparse==0.4.3
starlette==0.21.0
typing-extensions==4.4.0
tzdata==2022.5
uvicorn==0.20.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
Werkzeug==2.2.2
zipp==3.8.1
